refactor based on PR comments. Libarize-ContractGenerator. Follow a more static function style
