--- ContractPlugin/contract_plugin-csharp.csproj
@@ -1,25 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>net7.0</TargetFramework>
-        <RootNamespace>contract_plugin_csharp</RootNamespace>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <PublishTrimmed>true</PublishTrimmed>
-        <PublishReadyToRun>true</PublishReadyToRun>
-        <PublishSingleFile>true</PublishSingleFile>
-        <RuntimeIdentifier>osx-arm64</RuntimeIdentifier>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Google.Protobuf" Version="3.24.3"/>
-        <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.7.0"/>
-        <EditorConfigFiles Include="obj\Debug\net7.0\osx-arm64\contract_plugin-csharp.GeneratedMSBuildEditorConfig.editorconfig"/>
-    </ItemGroup>
-
-    <!-- Define a default build target -->
-    <Target Name="Build" Condition="'$(BuildingInsideVisualStudio)' != 'true'">
-        <MSBuild Projects="$(MSBuildProjectFile)" Targets="Restore;Build" Properties="Configuration=$(Configuration)"/>
-    </Target>
-
-</Project>
