--- ContractPlugin/contract_plugin-csharp.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.002.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "contract_plugin-csharp", "contract_plugin-csharp.csproj", "{B49755A2-A4FF-4BFF-98AC-6E22ABB64035}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B49755A2-A4FF-4BFF-98AC-6E22ABB64035}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B49755A2-A4FF-4BFF-98AC-6E22ABB64035}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B49755A2-A4FF-4BFF-98AC-6E22ABB64035}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B49755A2-A4FF-4BFF-98AC-6E22ABB64035}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {242D6504-F89E-4B05-BD57-ACE417AECE8E}
-	EndGlobalSection
-EndGlobal
